% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_refcats.R
\name{add_refcats}
\alias{add_refcats}
\title{add_refcats}
\usage{
add_refcats(df, model, ref_val = 1, sort_factors = FALSE)
}
\arguments{
\item{df}{Regression coefficients dataframe.}

\item{model}{Regression model.}

\item{ref_val}{Coefficient value for reference category.}

\item{sort_factors}{Option to sort factor variables in the table alphabetically.}
}
\value{
Regression model table with factor reference categories added.
}
\description{
Add reference categories for factor variables to a regression coefficients dataframe.
}
\examples{
\dontrun{
add_refcats(df, mod)
}
}
