% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbvmisc-package.R
\docType{package}
\name{bbvmisc-package}
\alias{bbvmisc-package}
\title{bbvmisc: BBV Team Miscellaneous Functions}
\description{
Useful functions used by the BBV team at Public Health Scotland and GCU.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alan-y/bbvmisc}
  \item Report bugs at \url{https://github.com/alan-y/bbvmisc/issues}
}

}
\author{
\strong{Maintainer}: Alan Yeung \email{alan.yeung@phs.scot} (\href{https://orcid.org/0000-0001-5226-3695}{ORCID})

Authors:
\itemize{
  \item Alan Yeung \email{alan.yeung@phs.scot} (\href{https://orcid.org/0000-0001-5226-3695}{ORCID})
}

Other contributors:
\itemize{
  \item Megan Glancy \email{megan.glancy2@phs.scot} (\href{https://orcid.org/0000-0003-2023-9353}{ORCID}) [contributor]
}

}
\keyword{internal}
