% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_p.R
\name{format_p}
\alias{format_p}
\title{format_p}
\usage{
format_p(x, cutoff = 0.001)
}
\arguments{
\item{x}{p-values.}

\item{cutoff}{Cut-off for p-values.}
}
\value{
Formatted p-values.
}
\description{
Format p-values.
}
\examples{
\dontrun{
format_p(x)
}
}
