% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftidy.R
\name{ftidy}
\alias{ftidy}
\title{ftidy}
\usage{
ftidy(model, exp = TRUE, drop_intercept = FALSE)
}
\arguments{
\item{model}{Regression model.}

\item{exp}{Option to exponentiate coefficients.}

\item{drop_intercept}{Option to drop intercept term from the output table.}
}
\value{
Regression model table with confidence intervals.
}
\description{
Fast version of the tidy function from the broom package that uses the normal approximation for confidence intervals.
}
\examples{
\dontrun{
ftidy(mod)
}
}
