% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_path}
\alias{get_path}
\title{format_p}
\usage{
get_path(x, server_path = "conf")
}
\arguments{
\item{x}{Input file path.}

\item{server_path}{Start of the server path.}
}
\value{
File path after desktop or server detection.
}
\description{
Sets the file path for whether or not the user is using R on the desktop or on the server.
}
\examples{
\dontrun{
get_path("//stats/my_folder")
}
}
