% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_univar.R
\name{glm_univar}
\alias{glm_univar}
\title{glm_univar}
\usage{
glm_univar(
  data,
  respvar,
  predvars,
  family = "poisson",
  exp = TRUE,
  additional_term = "",
  fast = TRUE
)
}
\arguments{
\item{data}{}

\item{respvar}{}

\item{predvars}{}

\item{family}{}

\item{exp}{}

\item{additional_term}{}

\item{fast}{}
}
\description{
Run a univariable `glm()` for a vector of predictor variables with tidy df output
}
