% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_list.R
\name{hb_list}
\alias{hb_list}
\title{hb_list}
\usage{
hb_list(scot = FALSE)
}
\arguments{
\item{scot}{Boolean. Include Scotland in the list.}
}
\value{
Character vector of NHS boards in Scotland.
}
\description{
List of NHS board names in Scotland. Helps to turn a healthboard
variable into a factor to ensure all boards are listed in output tables,
including those with zero counts.
}
\examples{
\dontrun{
hb_list()
}
}
