% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_recode.R
\name{hb_recode}
\alias{hb_recode}
\title{hb_recode}
\usage{
hb_recode(x, recode_to = "long", and_choice = "and")
}
\arguments{
\item{x}{NHS board vector.}

\item{recode_to}{Output format.}

\item{and_choice}{Choice for displaying and.}
}
\value{
NHS boards in desired form.
}
\description{
Recode Scottish NHS boards.
}
\examples{
\dontrun{
hb_recode(x)
}
}
