% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.R
\name{np}
\alias{np}
\title{np}
\usage{
np(n, p, digits = 0)
}
\arguments{
\item{n}{Counts.}

\item{p}{Proportions.}

\item{digits}{Number of digits to use after the decimal point in percentages.}
}
\value{
Counts with proportions in brackets.
}
\description{
Paste n and percentage values together.
}
\examples{
\dontrun{
np(n, p, digits = 1)
}
}
