% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_table.R
\name{pt_table}
\alias{pt_table}
\title{Title pt_table}
\usage{
pt_table(data, id_var = "bbv_id", outcome, vars, pt, var_col = TRUE)
}
\arguments{
\item{data}{Input dataset.}

\item{id_var}{unique identifier for individuals in dataset.}

\item{outcome}{binary outcome (outcome = 1).}

\item{vars}{Variables by which to group.}

\item{pt}{Variable containing person time.}

\item{var_col}{Option to include variable name as a separate column.}
}
\value{
Dataframe
}
\description{
Summarise number of people, number of events, and sum of person time by group.
}
\examples{
\dontrun{
pt_table(data, id_var="bbv_id", outcome="drd", vars=c("sex", "hb"), pt="person-years")
}

}
