% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex.R
\name{soundex}
\alias{soundex}
\title{soundex}
\usage{
soundex(name)
}
\arguments{
\item{name}{Name to get soundex code for.}
}
\value{
Soundex code of name.
}
\description{
Soundex code of name.
}
\examples{
\dontrun{
soundex("mysurname")
}
}
