% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_np_outcome.R
\name{table_np_outcome}
\alias{table_np_outcome}
\title{table_np_outcome}
\usage{
table_np_outcome(
  data,
  outcome,
  vars,
  digits = 0,
  np_only = TRUE,
  var_col = FALSE
)
}
\arguments{
\item{data}{Input dataset.}

\item{outcome}{Binary variable with outcome as "1".}

\item{vars}{Character vector of variables to get counts and percentages for.}

\item{digits}{Number of digits to use after the decimal point in percentages.}

\item{np_only}{Option to return only the combined count and percentage variable.}

\item{var_col}{Option to include variable name as a separate column.}
}
\value{
Data frame with counts and percentages of outcome for a list of variables.
}
\description{
Produce table of counts of outcome by variable, and percentage of full data within group.
}
\examples{
\dontrun{
table_np_outcome(data, outcome = "death", c("var1", "var2"))
}
}
