% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_pt.R
\name{table_pt}
\alias{table_pt}
\title{Title table_pt}
\usage{
table_pt(data, id_var = "bbv_id", outcome, vars, pt, var_col = TRUE)
}
\arguments{
\item{data}{Input dataset.}

\item{id_var}{unique identifier for individuals in dataset.}

\item{outcome}{binary outcome (outcome = 1).}

\item{vars}{Variables by which to group.}

\item{pt}{Variable containing person time.}

\item{var_col}{Option to include variable name as a separate column.}
}
\value{
Dataframe
}
\description{
Summarise number of people, number of events, and sum of person time by group.
}
\examples{
\dontrun{
table_pt(data, id_var="bbv_id", outcome="drd", vars=c("sex", "hb"), pt="person-years")
}

}
