% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ci.R
\name{x_ci}
\alias{x_ci}
\title{x_ci}
\usage{
x_ci(
  x,
  l,
  u,
  digits = 2,
  brackets = TRUE,
  ci_sep = "-",
  ci_prefix = NULL,
  na_vals = NULL
)
}
\arguments{
\item{x}{Central estimate.}

\item{l}{Lower estimate.}

\item{u}{Upper estimate.}

\item{digits}{Number of digits to use after decimal point.}

\item{brackets}{Option to enclose confidence interval in brackets.}

\item{ci_sep}{Separator for confidence interval.}

\item{ci_prefix}{Optional prefix to add text before confidence interval.}

\item{na_vals}{Value to use to replace missing values; if NULL, uses x.}
}
\value{
Estimate combined with confidence interval as a string.
}
\description{
Combine estimate with confidence interval into one string.
}
\examples{
\dontrun{
x_ci(1.2, 0.7, 1.5, ci_sep = "--", ci_prefix = "95\% CI ")
}
}
