% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_multivar.R
\name{glm_multivar}
\alias{glm_multivar}
\title{glm_multivar}
\usage{
glm_multivar(
  data,
  respvar,
  predvars,
  family = "poisson",
  exp = TRUE,
  additional_term = "",
  fast = TRUE
)
}
\arguments{
\item{data}{}

\item{respvar}{}

\item{predvars}{}

\item{family}{}

\item{exp}{}

\item{additional_term}{}
}
\description{
Run a multivariable `glm()` for a vector of predictor variables with tidy df output
}
