% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_univar.R
\name{glm_univar}
\alias{glm_univar}
\title{glm_univar}
\usage{
glm_univar(
  data,
  respvar,
  predvars,
  family = "poisson",
  exp = TRUE,
  additional_term = "",
  fast = TRUE
)
}
\arguments{
\item{data}{Input dataset}

\item{respvar}{Single character string of response variable}

\item{predvars}{Vector of character strings of predictor variables}

\item{family}{Single character string of family function}

\item{exp}{Option to exponentiate model estimates (default is TRUE)}

\item{additional_term}{Option to add additional terms to model (default is blank)}

\item{fast}{Option to use normal approximation for confidence intervals (default is TRUE)}
}
\value{
Regression model table with confidence intervals
}
\description{
Run a univariable `glm()` for a vector of predictor variables with tidy df output
}
\examples{
\dontrun{
glm_univar(dat, "var", c("var1", "var2"), additional_term="+ offset(log(py))")
}
}
