% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.R
\name{np}
\alias{np}
\title{np}
\usage{
np(n, p, acc = 1)
}
\arguments{
\item{n}{Counts.}

\item{p}{Proportions.}

\item{acc}{Number to round to for percentages in the combined count and percentage variable (e.g. 0.1 shows one decimal place for the percentage).}
}
\value{
Counts with proportions in brackets.
}
\description{
Paste n and percentage values together.
}
\examples{
\dontrun{
np(n, p, acc = 0.1)
}
}
