% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_table.R
\name{np_table}
\alias{np_table}
\title{np_table}
\usage{
np_table(data, vars, np_only = TRUE)
}
\arguments{
\item{data}{Input dataset.}

\item{vars}{Character vector of variables to get counts and percentages for.}

\item{np_only}{Option to return only the combined count and percentage variable.}
}
\value{
Data frame with counts and percentages for a list of variables.
}
\description{
Fast version of the tidy function from the broom package that uses the normal approximation for confidence intervals.
}
\examples{
\dontrun{
np_table(dat, c("var1", "var2"))
}
}
