% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_table.R
\name{np_table}
\alias{np_table}
\title{np_table}
\usage{
np_table(
  data,
  vars,
  digits = 0,
  comma = FALSE,
  np_only = TRUE,
  header = TRUE,
  var_col = FALSE
)
}
\arguments{
\item{data}{Input dataset.}

\item{vars}{Character vector of variables to get counts and percentages for.}

\item{digits}{Number of digits to use after the decimal point in percentages.}

\item{comma}{Option to use comma separator on thousands.}

\item{np_only}{Option to return only the combined count and percentage variable.}

\item{header}{Option to use variable names as headers.}

\item{var_col}{Option to include variable name as a separate column.}
}
\value{
Data frame with counts and percentages for a list of variables.
}
\description{
Produce table of counts and percentages for a list of variables in a dataset.
}
\examples{
\dontrun{
np_table(dat, c("var1", "var2"), digits = 1)
}
}
