% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_byar_ci.R
\name{pois_byar_ci}
\alias{pois_byar_ci}
\title{Title pois_byar_ci}
\usage{
pois_byar_ci(data, x, pt, factor = 1000, conf.level = 0.95)
}
\arguments{
\item{data}{Input dataset containing counts and person time}

\item{x}{count or vector of counts}

\item{pt}{person-time at risk or vector of person-times}

\item{factor}{per how many person-time}

\item{conf.level}{confidence level (default = 0.95)}
}
\value{
original dataframe with rate, lower, upper, and rate_ci variables added
}
\description{
Calculate confidence intervals for Poisson counts or rates using Byar's method, and return tidy output
}
\examples{
\dontrun{
pois_byar_ci(data=counts, x="deaths", pt="person_years")
 }
}
